+++ src/api/MonsterOfTheWeek.Api/Repositories/IReferenceRepository.cs
@@ -5,12 +5,14 @@ namespace MonsterOfTheWeek.Api.Repositories;
 public interface IReferenceRepository
 {
     Task<IReadOnlyList<AdventureType>> GetAdventureTypesAsync(CancellationToken cancellationToken);
+    Task<IReadOnlyList<MonsterArchetype>> GetMonsterArchetypesAsync(CancellationToken cancellationToken);
     Task<IReadOnlyList<MonsterType>> GetMonsterTypesAsync(CancellationToken cancellationToken);
     Task<IReadOnlyList<MinionType>> GetMinionTypesAsync(CancellationToken cancellationToken);
     Task<IReadOnlyList<LocationType>> GetLocationTypesAsync(CancellationToken cancellationToken);
     Task<IReadOnlyList<BystanderType>> GetBystanderTypesAsync(CancellationToken cancellationToken);
     Task<IReadOnlyList<WeaponTag>> GetWeaponTagsAsync(CancellationToken cancellationToken);
     Task AddAdventureTypeAsync(AdventureType adventureType, CancellationToken cancellationToken);
+    Task AddMonsterArchetypeAsync(MonsterArchetype monsterArchetype, CancellationToken cancellationToken);
     Task AddMonsterTypeAsync(MonsterType monsterType, CancellationToken cancellationToken);
     Task AddMinionTypeAsync(MinionType minionType, CancellationToken cancellationToken);
     Task AddLocationTypeAsync(LocationType locationType, CancellationToken cancellationToken);