+++ src/api/MonsterOfTheWeek.Api/Repositories/ReferenceRepository.cs
@@ -9,6 +9,9 @@ public sealed class ReferenceRepository(MotwDbContext dbContext) : IReferenceRep
     public async Task<IReadOnlyList<AdventureType>> GetAdventureTypesAsync(CancellationToken cancellationToken) =>
         await dbContext.AdventureTypes.AsNoTracking().OrderBy(x => x.Name).ToListAsync(cancellationToken);
 
+    public async Task<IReadOnlyList<MonsterArchetype>> GetMonsterArchetypesAsync(CancellationToken cancellationToken) =>
+        await dbContext.MonsterArchetypes.AsNoTracking().OrderBy(x => x.Name).ToListAsync(cancellationToken);
+
     public async Task<IReadOnlyList<MonsterType>> GetMonsterTypesAsync(CancellationToken cancellationToken) =>
         await dbContext.MonsterTypes.AsNoTracking().OrderBy(x => x.Name).ToListAsync(cancellationToken);
 
@@ -26,6 +29,9 @@ public sealed class ReferenceRepository(MotwDbContext dbContext) : IReferenceRep
 
     public Task AddAdventureTypeAsync(AdventureType adventureType, CancellationToken cancellationToken) =>
         dbContext.AdventureTypes.AddAsync(adventureType, cancellationToken).AsTask();
+
+    public Task AddMonsterArchetypeAsync(MonsterArchetype monsterArchetype, CancellationToken cancellationToken) =>
+        dbContext.MonsterArchetypes.AddAsync(monsterArchetype, cancellationToken).AsTask();
 
     public Task AddMonsterTypeAsync(MonsterType monsterType, CancellationToken cancellationToken) =>
         dbContext.MonsterTypes.AddAsync(monsterType, cancellationToken).AsTask();